build flags ftw - linux,bsd,sol,osx,win
