+++ internal/agent/main.go
@@ -7,13 +7,9 @@ package agent
 
 import (
 	"context"
-	"fmt"
 	"os"
 	"os/signal"
-	"runtime"
-	"syscall"
 
-	"github.com/alecthomas/units"
 	"github.com/circonus-labs/circonus-agent/internal/plugins"
 	"github.com/circonus-labs/circonus-agent/internal/release"
 	"github.com/circonus-labs/circonus-agent/internal/reverse"
@@ -43,7 +39,7 @@ func New() (*Agent, error) {
 	}
 
 	// Handle shutdown via a.shutdownCtx
-	signal.Notify(a.signalCh, os.Interrupt, syscall.SIGTERM, syscall.SIGHUP, syscall.SIGPIPE, syscall.SIGTRAP)
+	signalNotifySetup(a.signalCh)
 
 	a.shutdownCtx, a.shutdown = context.WithCancel(context.Background())
 
@@ -115,35 +111,6 @@ func (a *Agent) Wait() error {
 	}
 
 	return nil
-}
-
-// handleSignals runs the signal handler thread
-func (a *Agent) handleSignals() {
-	const stacktraceBufSize = 1 * units.MiB
-
-	// pre-allocate a buffer
-	buf := make([]byte, stacktraceBufSize)
-
-	for {
-		select {
-		case <-a.shutdownCtx.Done():
-			log.Debug().Msg("Shutting down")
-			return
-		case sig := <-a.signalCh:
-			log.Info().Str("signal", sig.String()).Msg("Received signal")
-			switch sig {
-			case os.Interrupt, syscall.SIGTERM:
-				a.shutdown()
-			case syscall.SIGPIPE, syscall.SIGHUP:
-				// Noop
-			case syscall.SIGTRAP:
-				stacklen := runtime.Stack(buf, true)
-				fmt.Printf("=== received SIGINFO ===\n*** goroutine dump...\n%s\n*** end\n", buf[:stacklen])
-			default:
-				panic(fmt.Sprintf("unsupported signal: %v", sig))
-			}
-		}
-	}
 }
 
 // stopSignalHandler disables the signal handler
